Fix container files ATS review feedback
Use generated Java and Python container file option shapes in polyglot apphosts. Make numeric container file options tolerant of guest SDK decimal-form numbers by validating and coercing finite integral doubles before invoking the underlying container file API.

+++ src/Aspire.Hosting/Ats/ContainerFilesOptions.cs
@@ -9,19 +9,22 @@ namespace Aspire.Hosting.Ats;
 [AspireDto]
 internal sealed class ContainerFilesOptions
 {
+    // Guest SDKs commonly represent DTO numbers as floating-point values. The export
+    // validates and coerces these to the integral owner, group, and mode values.
+
     /// <summary>
     /// The default owner UID for the created or updated file system entries. Defaults to 0 for root if not set.
     /// </summary>
-    public int? DefaultOwner { get; init; }
+    public double? DefaultOwner { get; init; }
 
     /// <summary>
     /// The default group ID for the created or updated file system entries. Defaults to 0 for root if not set.
     /// </summary>
-    public int? DefaultGroup { get; init; }
+    public double? DefaultGroup { get; init; }
 
     /// <summary>
     /// The Unix umask to apply to files or directories without explicit permissions.
     /// Use octal literals in JavaScript or TypeScript, for example <c>0o022</c>.
     /// </summary>
-    public int? Umask { get; init; }
+    public double? Umask { get; init; }
 }
